% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isotopes.R
\name{calc_isotope_distribution}
\alias{calc_isotope_distribution}
\title{Compute the isotope envelope for a peptide sequence}
\usage{
calc_isotope_distribution(sequence, n_isotopes = 5L)
}
\arguments{
\item{sequence}{character(1). Single-letter amino acid sequence.}

\item{n_isotopes}{integer(1). Number of isotope peaks to return (M+0 to
M+\code{n_isotopes}). Default 5 (i.e. M+0 through M+4).}
}
\value{
data.frame with columns:
  \describe{
    \item{isotope}{integer. Offset from monoisotopic peak (0, 1, 2, ...).}
    \item{mz_offset}{numeric. Mass offset in Da (multiples of neutron mass).}
    \item{probability}{numeric. Relative abundance (normalised to sum = 1).}
    \item{rel_intensity}{numeric. Intensity as a percentage of the most
      abundant peak (0-100).}
  }
}
\description{
Estimates the isotope abundance distribution for a peptide
  using the averagine approximation (Senko et al. 1995, J. Am. Soc. Mass
  Spectrom. 6:229-233). The elemental composition is approximated by
  scaling the averagine monomer to the number of residues and convolving
  element-wise Poisson distributions.
}
\examples{
calc_isotope_distribution("EVQLVESGGG")
calc_isotope_distribution("PEPTIDEPEPTIDEPEPTIDE", n_isotopes = 4)
}
\references{
Senko et al. (1995) \emph{J. Am. Soc. Mass Spectrom.} 6:229.
}
