% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adc_design.R
\name{plot_deconjugation}
\alias{plot_deconjugation}
\title{Plot ADC deconjugation kinetics curve}
\usage{
plot_deconjugation(result)
}
\arguments{
\item{result}{list. Output of \code{\link{predict_deconjugation}}.}
}
\value{
A \code{ggplot2} object. Print or save with \code{ggsave()}.
}
\description{
Creates a ggplot2 line plot of DAR over time from the output
  of \code{\link{predict_deconjugation}}, with a shaded sub-therapeutic
  zone and dashed threshold line.
}
\examples{
\dontrun{
  res <- predict_deconjugation("Maleimide-thiol (MC-VC-PABC / SMCC)",
                               site = "Cys-engineered")
  plot_deconjugation(res)
}
}
