% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/digest.R
\name{PROTON_MASS}
\alias{PROTON_MASS}
\title{Mass of a proton (Da)}
\format{
A numeric scalar.
}
\usage{
PROTON_MASS
}
\description{
Monoisotopic mass of a proton (H+), used to convert neutral
  masses to m/z values for protonated ions.
}
