% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/digest.R
\name{WATER_MASS}
\alias{WATER_MASS}
\title{Mass of water (Da)}
\format{
A numeric scalar.
}
\usage{
WATER_MASS
}
\description{
Monoisotopic mass of water (H2O), used to compute peptide
  neutral masses and y-ion series.
}
